feat(loaders): created models and data loaders
Created loaders dir. This module contains the models and the data
loader modules.

The models_loader module is in charge of load the models from MongoDB.
If there are not saved models, it creates, fits and saves them.

The data loader module is in charge of download the data JSON.